python/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "dorea-inference"
version = "0.1.0"
description = "Dorea inference subprocess — RAUNE-Net + Depth Anything V2"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "transformers>=4.38",
    "Pillow>=10.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = ["pytest>=7"]

[tool.setuptools.packages.find]
where = ["."]
include = ["dorea_inference*"]
